tiktok_downloader/video_parser/requirements.txt
opencv-python-headless
openai-whisper
transformers
torch
pytesseract
easyocr
pandas